refactor(carousel): modularize types and add tests
- Removed `types.ts` from the carousel component and split type definitions into corresponding locations:
  - Core types moved to `@src/hooks/components/types.ts`.
  - Carousel-specific types moved to `@src/components/carousel/components/types.ts`.
- Updated imports across various carousel components to use the new type files.
- Enhanced testing:
  - Added snapshot tests for `CarouselArrowIndex` and `CarouselArrows` components.
  - Expanded test support by mocking `window.matchMedia` in `setupTest.ts`.
- Minor adjustments to component files for better resilience and modularity.
