+++ setupTest.ts
@@ -5,3 +5,13 @@ expect.extend(matchers);
 
 // Mocks
 vi.mock('react-i18next');
+
+
+// @ts-ignore
+window.matchMedia = window.matchMedia || function() {
+  return {
+    matches : false,
+    addListener : function() {},
+    removeListener: function() {}
+  };
+};
