+++ src/components/carousel/components/test/CarouselArrow.test.tsx
@@ -1,0 +1,11 @@
+import { describe, it, expect } from 'vitest';
+import { render } from '@testing-library/react';
+import {CarouselArrows} from "@src/components/carousel";
+
+
+describe('CarouselArrows component testing', () => {
+  it('to match snapshot', () => {
+    const{baseElement} = render(<CarouselArrows onNext={() => {}} onPrev={() => {}} />);
+    expect(baseElement).toMatchSnapshot();
+  });
+})
