+++ src/components/iconify/iconify.tsx
@@ -14,6 +14,7 @@ interface Props extends BoxProps {
 
 const Iconify = forwardRef<SVGElement, Props>(({ icon, width = 20, sx, ...other }, ref) => (
   <Box
+    data-testid="iconify"
     ref={ref}
     component={Icon}
     className="component-iconify"
