+++ src/hooks/components/types.ts
@@ -1,0 +1,18 @@
+import React from "react";
+import Carousel, {Settings} from "react-slick";
+
+export type CarouselReturnType = {
+  currentIndex: number;
+  nav: Carousel | undefined;
+  carouselSettings: Settings;
+  carouselRef: React.MutableRefObject<Carousel | null>;
+  //
+  onPrev: VoidFunction;
+  onNext: VoidFunction;
+  onSetNav: VoidFunction;
+  onTogo: (index: number) => void;
+  //
+  setNav: React.Dispatch<React.SetStateAction<Carousel | undefined>>;
+  setCurrentIndex: React.Dispatch<React.SetStateAction<number>>;
+};
+
